// frontend/src/api/generated/models/cashAdvanceResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface CashAdvanceResponse {
  id: string;
  operator_id: string;
  operator_name: string;
  date: string;
  amount: string;
  remaining_amount: string;
  is_paid: boolean;
  notes: string | null;
}